suggestion follow api v2
refractor suggestion follow


api suggestUserV2


api suggest user


update variable


remove unused variable
